% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefVar.R
\name{CoefVar}
\alias{CoefVar}
\title{R6 Coefficient of Variation (cv)}
\arguments{
\item{x}{An \code{R} object. Currently there are methods for numeric vectors}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}

\item{digits}{integer indicating the number of decimal places to be used.}
}
\description{
The R6 class \code{CoefVar} for the coefficient of variation
  (cv)
}
\details{
\describe{ \item{\strong{Coefficient of Variation}}{ The \emph{cv}
  is a measure of relative dispersion representing the degree of variability
  relative to the mean \code{[1]}. Since \eqn{cv} is unitless, it is useful
  for comparison of variables with different units. It is also a measure of
  homogeneity \code{[1]}. } }
}
\examples{
x <- c(
    0.2, 0.5, 1.1, 1.4, 1.8, 2.3, 2.5, 2.7, 3.5, 4.4,
    4.6, 5.4, 5.4, 5.7, 5.8, 5.9, 6.0, 6.6, 7.1, 7.9
)
CoefVar$new(x)$est()
cv_x <- CoefVar$new(x, digits = 2)
cv_x$est()
cv_x$est_corr()
R6::is.R6(cv_x)
}
\references{
\code{[1]} Albatineh, AN., Kibria, BM., Wilcox, ML., & Zogheib,
  B, 2014, Confidence interval estimation for the population coefficient of
  variation using ranked set sampling: A simulation study, Journal of Applied
  Statistics, 41(4), 733–751, DOI:
  \href{http://doi.org/10.1080/02664763.2013.847405}{http://doi.org/10.1080/02664763.2013.847405}
}
