% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest_length.R
\name{shortest_length}
\alias{shortest_length}
\title{shortest-length confidence interval for cv}
\description{
The data frame including the values of a and b for the
             shortest-length confidence interval for the coefficient of
             variation (cv)
}
\references{
Panichkitkosolkul, W., 2013, Confidence Intervals for the
                Coefficient of Variation in a Normal Distribution with a
                Known Population Mean, Journal of Probability and Statistics,
                2013, 1–11, \href{http://doi.org/10.1155/2013/324940}{http://doi.org/10.1155/2013/324940}
}
