% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{x}
\alias{x}
\title{The input numeric vector x example}
\description{
The input numeric vector x example
}
\references{
[1] Bonett, DG., 2006, Confidence interval for a coefficient of
                quartile variation, Computational Statistics & Data Analysis,
                50(11), 2953-7, DOI: \href{http://doi.org/10.1016/j.csda.2005.05.007}{http://doi.org/10.1016/j.csda.2005.05.007}
}
